+++ Plainquire.Sort.Tests/Plainquire.Sort.Tests/Services/EntitySortFunctions.cs
@@ -1,5 +1,4 @@
 ﻿using Newtonsoft.Json;
-using Plainquire.Sort.Abstractions;
 using Plainquire.Sort.Newtonsoft;
 using Plainquire.Sort.Tests.Models;
 using System;
@@ -10,7 +9,7 @@ using System.Reflection;
 
 namespace Plainquire.Sort.Tests.Services;
 
-public delegate List<TEntity> EntitySortFunction<TEntity>(IEnumerable<TEntity> items, EntitySort<TEntity> sort, SortConfiguration? configuration = null, ISortInterceptor? interceptor = null);
+public delegate List<TEntity> EntitySortFunction<TEntity>(IEnumerable<TEntity> items, EntitySort<TEntity> sort, ISortInterceptor? interceptor = null);
 
 [ExcludeFromCodeCoverage]
 public static class EntitySortFunctions
@@ -31,26 +30,26 @@ public static class EntitySortFunctions
             .MakeGenericMethod(entityType)
             .Invoke(null, [])!;
 
-    private static List<TEntity> SortDirectByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
-        => testItems.OrderBy(sort, configuration, interceptor).ToList();
+    private static List<TEntity> SortDirectByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
+        => testItems.OrderBy(sort, interceptor).ToList();
 
-    private static List<TEntity> SortNetCloneByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
-        => testItems.OrderBy(sort.Clone(), configuration, interceptor).ToList();
+    private static List<TEntity> SortNetCloneByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
+        => testItems.OrderBy(sort.Clone(), interceptor).ToList();
 
-    private static List<TEntity> SortNewtonCloneByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
-        => testItems.OrderBy(sort.NewtonsoftClone(), configuration, interceptor).ToList();
+    private static List<TEntity> SortNewtonCloneByLinq<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
+        => testItems.OrderBy(sort.NewtonsoftClone(), interceptor).ToList();
 
-    private static List<TEntity> SortDirectByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
+    private static List<TEntity> SortDirectByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
         where TEntity : class
-        => testItems.OrderByEF(sort, configuration, interceptor);
+        => testItems.OrderByEF(sort, interceptor);
 
-    private static List<TEntity> SortNetCloneByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
+    private static List<TEntity> SortNetCloneByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
         where TEntity : class
-        => testItems.OrderByEF(sort.Clone(), configuration, interceptor);
+        => testItems.OrderByEF(sort.Clone(), interceptor);
 
-    private static List<TEntity> SortNewtonCloneByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
+    private static List<TEntity> SortNewtonCloneByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
         where TEntity : class
-        => testItems.OrderByEF(sort.NewtonsoftClone(), configuration, interceptor);
+        => testItems.OrderByEF(sort.NewtonsoftClone(), interceptor);
 
     private static EntitySort<TEntity> NewtonsoftClone<TEntity>(this EntitySort<TEntity> sort)
     {
@@ -59,13 +58,13 @@ public static class EntitySortFunctions
         return JsonConvert.DeserializeObject<EntitySort<TEntity>>(json, serializerSettings)!;
     }
 
-    private static List<TEntity> OrderByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, SortConfiguration? configuration, ISortInterceptor? interceptor)
+    private static List<TEntity> OrderByEF<TEntity>(this IEnumerable<TEntity> testItems, EntitySort<TEntity> sort, ISortInterceptor? interceptor)
         where TEntity : class
     {
         using var dbContext = new TestDbContext<TEntity>();
         dbContext.Set<TEntity>().AddRange(testItems);
         dbContext.SaveChanges();
 
-        return dbContext.Set<TEntity>().OrderBy(sort, configuration, interceptor).ToList();
+        return dbContext.Set<TEntity>().OrderBy(sort, interceptor).ToList();
     }
 }