+++ Plainquire.Sort/Plainquire.Sort/Interfaces/ISortInterceptor.cs
@@ -1,5 +1,4 @@
-﻿using Plainquire.Sort.Abstractions;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 
 namespace Plainquire.Sort;
@@ -16,10 +15,9 @@ public interface ISortInterceptor
     /// <typeparam name="TEntity">The type of the sorted entity.</typeparam>
     /// <param name="source">A sequence of values to sort.</param>
     /// <param name="sort">The sort order.</param>
-    /// <param name="configuration">The active sort order configuration</param>
     /// <returns>An <see cref="IOrderedEnumerable{TEntity}"/> with applied sort order or <c>null</c> to use the default sort order.</returns>
     [SuppressMessage("ReSharper", "ParameterTypeCanBeEnumerable.Global", Justification = "Required for functionallity.")]
-    public IOrderedQueryable<TEntity>? OrderBy<TEntity>(IQueryable<TEntity> source, PropertySort sort, SortConfiguration configuration);
+    public IOrderedQueryable<TEntity>? OrderBy<TEntity>(IQueryable<TEntity> source, PropertySort sort);
 
     /// <summary>
     /// Performs a subsequent ordering of a sequence according to the given <paramref name="sort"/>.
@@ -27,8 +25,7 @@ public interface ISortInterceptor
     /// <typeparam name="TEntity">The type of the sorted entity.</typeparam>
     /// <param name="source">A sequence of values to sort.</param>
     /// <param name="sort">The sort order.</param>
-    /// <param name="configuration">The active sort order configuration</param>
     /// <returns>An <see cref="IOrderedEnumerable{TEntity}"/> with applied sort order or <c>null</c> to use the default sort order.</returns>
     [SuppressMessage("ReSharper", "ParameterTypeCanBeEnumerable.Global", Justification = "Required for functionallity.")]
-    public IOrderedQueryable<TEntity>? ThenBy<TEntity>(IOrderedQueryable<TEntity> source, PropertySort sort, SortConfiguration configuration);
+    public IOrderedQueryable<TEntity>? ThenBy<TEntity>(IOrderedQueryable<TEntity> source, PropertySort sort);
 }